store accounts to persistent DB
